+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2025-07-22
+
+- Add `convertToJpegAsync` method
+- Add `getRgba` method
+- Add `ConvertToJpegIsolateMixin` mixin
+- Add `convertToJpegAsync` method
+
 ## [1.0.1] - 2025-06-26
 
 - call `image.dispose()`